pyproject.toml
[project]
name = "jugg-finder"
version = "0.1.0"
description = "Local freelance lead generator for Boise, Idaho — finds small businesses with poor web presence."
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tenacity>=9.1.4",
    "uvicorn>=0.45.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/scorer/ai.py" = ["E501"]  # Prompt template contains long lines by design

[tool.pytest.ini_options]
asyncio_mode = "auto"
